+++ src/main/java/org/dynjs/runtime/builtins/types/array/prototype/Splice.java
@@ -20,20 +20,20 @@ public class Splice extends AbstractNativeFunction {
         // 15.4.4.11
         JSObject o = Types.toObject(context, self);
 
-        int len = Types.toInteger(context, o.get(context, "length"));
-        int relativeStart = Types.toInteger(context, args[0]);
-        int actualStart = relativeStart;
+        long len = Types.toInteger(context, o.get(context, "length"));
+        long relativeStart = Types.toInteger(context, args[0]);
+        long actualStart = relativeStart;
 
         if (relativeStart < 0) {
             actualStart = len + relativeStart;
         }
 
-        int deleteCount = Types.toInteger(context, args[1]);
-        int actualDeleteCount = Math.min(Math.max(deleteCount, 0), len - actualStart);
+        long deleteCount = Types.toInteger(context, args[1]);
+        long actualDeleteCount = Math.min(Math.max(deleteCount, 0), len - actualStart);
 
         DynArray a = BuiltinArray.newArray(context);
 
-        for (int k = 0; k < actualDeleteCount; ++k) {
+        for (long k = 0; k < actualDeleteCount; ++k) {
             if (o.hasProperty(context, "" + (actualStart + k))) {
                 final Object fromValue = o.get(context, "" + (actualStart + k));
                 a.defineOwnProperty(context, "" + k, new PropertyDescriptor() {
@@ -47,10 +47,10 @@ public class Splice extends AbstractNativeFunction {
             }
         }
 
-        int itemCount = args.length - 2;
+        long itemCount = args.length - 2;
 
         if (itemCount < actualDeleteCount) {
-            for (int k = actualStart; k < (len - actualDeleteCount); ++k) {
+            for (long k = actualStart; k < (len - actualDeleteCount); ++k) {
                 if (o.hasProperty(context, "" + (k + actualDeleteCount))) {
                     final Object fromValue = o.get(context, "" + (k + actualDeleteCount));
                     o.put(context, "" + (k + itemCount), fromValue, true);
@@ -59,11 +59,11 @@ public class Splice extends AbstractNativeFunction {
                 }
             }
 
-            for (int k = len; k > (len - actualDeleteCount + itemCount); --k) {
+            for (long k = len; k > (len - actualDeleteCount + itemCount); --k) {
                 o.delete(context, "" + (k - 1), true);
             }
         } else if (itemCount > actualDeleteCount) {
-            for (int k = (len - actualDeleteCount); k > actualStart; --k) {
+            for (long k = (len - actualDeleteCount); k > actualStart; --k) {
                 if (o.hasProperty(context, "" + (k + actualDeleteCount - 1))) {
                     final Object fromValue = o.get(context, "" + (k + actualDeleteCount - 1));
                     o.put(context, "" + (k + itemCount - 1), fromValue, true);
@@ -73,8 +73,8 @@ public class Splice extends AbstractNativeFunction {
             }
         }
 
-        for (int k = actualStart, i = 0; i < itemCount; ++k, ++i) {
-            o.put(context, "" + k, args[i + 2], true);
+        for (long k = actualStart, i = 0; i < itemCount; ++k, ++i) {
+            o.put(context, "" + k, args[ (int) i + 2], true);
         }
 
         return a;
